[Element Tracker] Support adding custom event callback for specific ID.
Because of the way custom events are implemented, the previous callback
did not filter by element id, so you'd get a callback for the same
event regardless of which element (or element in the specified context)
generated it.

However, with some of the Webium use cases, the same event might be
generated by a number of different elements, and the caller might only
care about one.

This CL adds filtered versions of add custom event callback that will
only call the callback if the source element has a specified id.

Change-Id: I5570788d7e0eadfbcb67cd9243a81a40b1fe1a46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7170761
Reviewed-by: Keren Zhu <kerenzhu@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547772}