[dbsc] Add DeleteAllSigningKeysSlowly to StatefulUnexportableKeyProvider.
This change introduces a new method, `DeleteAllSigningKeysSlowly`, to
the `StatefulUnexportableKeyProvider` interface. This method allows for
the deletion of all signing keys matching the provider's configuration.
The macOS implementation is currently a stub, and the mock and fake
implementations are updated to include this new method.

Bug: 455539044
Change-Id: I0c7d657e2f91825c83712ae5c58512948e31e05c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7159235
Commit-Queue: Jan Wilken Dörrie <jdoerrie@chromium.org>
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547783}