[Persist] Change storage ID to a 128-bit token
Because we restore per window/OTR state using a `next_storage_id_` in
TabStateStorageService is not viable as it will not "advance" to the
next ID correctly unless every tab is loaded at startup all at once.

Using an ID allocated by the DB leads to issues as it imposes hard
sequencing requirements between the order of tabs and different
collection types when stored. The children and payload protos may
contain what are effectively "foreign keys" and if the tabs or other
collections referenced by them don't get written first we can't convert
to the IDs correctly.

This leads to the best solution being to use UUIDs. base::Uuid uses a
string though so it is best to use base::Token or base::UnguessableToken
instead to improve performance. The only concern would be collisions
leading to a corruption of the database, but this would be exceedingly
unlikely.

Fixed: 462096869
Change-Id: I92ce2711f07e9261ce03dc0d3359eb9d373e6afb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7169919
Reviewed-by: Sky Malice <skym@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547792}