[aim][popup] Fix popup opening after adding image/file on linux.
The flow on Linux differs from other platforms in that on closing the
file dialog focus seems to be restored to the omnibox which causes the
AIM popup to close before it can open and get focus. What seems to
happen is the following:

1. The file selector dialog is presented, causing an activation change
which hides the popup.

2. The user selects a file and the file selector triggers a presentation
of the AIM popup.

3. Sometimes (I see this most often if the composebox textbox has no
focus) right after the popup presents a Widget activation change happens
(perhaps due to the focus being restored to the ommibox?).

4. The widget activation change triggers the closure of the popup.

This change stops Widget activation-related visibility changes while the
popup is in a transition state (the 100ms duration after the popup state
changes to kAim). This stops the above flow after step 2 and allows the
popup to show.

Bug: b:462132809
Change-Id: I7dc755e8be6dc53a318824ffeb74172a38cf0871
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7171478
Auto-Submit: Aviv Kiss <avivkiss@google.com>
Commit-Queue: Aviv Kiss <avivkiss@google.com>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547819}