Add JS optimization chip to omnibar
This CL displays a chip in the omnibar when v8
optimizations are disabled in the WebContents.

Currently, the chip just displays a tooltip.
In future work (crbug.com/458711797) we will show a bubble
that allows the user to enable v8 optimizations.

Also, currently the chip uses the code icon (< >) but
it will be updated to use the v8 icon in future work
(crbug.com/457422266).

Bug: 458714131
Change-Id: I24b8f162d39795608b7268c328b33113d41f8965
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7132103
Reviewed-by: Kaan Alsan <alsan@chromium.org>
Reviewed-by: Javier Castro <jacastro@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Commit-Queue: Adam Abeshouse <aabeshouse@google.com>
Cr-Commit-Position: refs/heads/main@{#1547851}