Add tests for Pointer Lock release conditions.
This change adds unit tests to verify the behavior of pointer lock release in WindowAndroid. It checks that the native side is notified and capture is released when onPointerCaptureChange is called with false or when the focus is lost.

Bug: b:462351889
Change-Id: I74400bc9c61d658b1f06d3a0350c3676582c15ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7176786
Auto-Submit: Abdelrahman Eed <abdoeed@google.com>
Reviewed-by: Bo Liu <boliu@chromium.org>
Commit-Queue: Abdelrahman Eed <abdoeed@google.com>
Commit-Queue: Bo Liu <boliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547864}