[iOS] Add snapshot spinner when the tab is loading
This CL adds a spinner to the tab grid cells to indicate that a tab is
loading. The spinner is shown when the tab's web content is being
realized, providing immediate feedback to the user and making the UI
feel more responsive.

This change includes:
- Adding `{show,hide}{Favicon,Snapshot}ActivityIndicator` properties to `GridCell`.
- Updating `BaseGridViewController` to show/hide the activity indicator based on the tab's loading state.

See video:
https://drive.google.com/file/d/1xoUPn5SlZBLOmCCsDtcrN-t-FmktIRCX/view?usp=sharing

Fixed: 457967291
Change-Id: Ic2baaac8bac3cd22015198c0deb4fdeaad822b08
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7173258
Commit-Queue: Aliona Dangla <alionadangla@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547874}