[dbsc] Add unexportable keys proxy service
Chrome's DBSC implementation lives in the network process. The lowered
privileges that it has constrain the use of Unexportable Key Service on
certain platforms, like macOS for example.

To support DBSC in macOS we thus need to change how the DBSC has access
to the UKS. Instead of directly using the UKS implementation in the
network process, we will move it to the browser process. And from the
browser service, we will expose a new proxy mojo service which the
network process can use.

This proxy service will implement all the asynchronous methods of the
UKS: key generation, key loading, and signing. However, the UKS also
has some synchronous methods which can't be forwarded to an async mojo
method. To solve this issue, the methods in the mojo service will also
return additional information that can be cached in the network process
and used to implement the synchronous part of the UKS interface.

So for example when creating or generating keys the mojo proxy won't
simply return the new key, but also metadata about what algorithm it is
using.

This CL starts by adding the GenerateKey method, with the rest to follow
in future CLs.

NO_IFTTT=file is added now and IFTTT is misfiring

Bug: 452303497, 443932320, 452300505
Change-Id: I5feb9cd1b71f8ec52c36846a035bc7775ec066c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7075194
Commit-Queue: Konstantin Andrikopoulos <andrikopoulos@google.com>
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547877}