[rustmojo] Add tests for mojom unions
This CL adds testing for the parsing and code generation of mojom
unions. Since unions are a new type of structured data (we previously
only had structs), this required tweaking our testing format somewhat
to accommodate.

The tests are more elaborate than for some other types because unions
are extremely complicated, so I want to be sure we have as much
coverage as possible. In particular, I have manually verified that the
last test produces the same wire bytes as the C++ code.

Bug: 444509367
Change-Id: Id0a8bfcda3ebc4e8e01888de8ce7af87dd2ed76b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7158573
Commit-Queue: Devon Loehr <dloehr@google.com>
Reviewed-by: Julia Hansbrough <flowerhack@google.com>
Cr-Commit-Position: refs/heads/main@{#1547893}