[dbsc] Add Sign in proxy service
Chrome's DBSC implementation lives in the network process. The lowered
privileges that it has constrain the use of Unexportable Key Service on
certain platforms, like macOS for example.

To support DBSC in macOS we thus need to change how the DBSC has access
to the UKS. Instead of directly using the UKS implementation in the
network process, we will move it to the browser process. And from the
browser service, we will expose a new proxy mojo service which the
network process can use.

This proxy service will implement all the asynchronous methods of the
UKS: key generation, key loading, and signing. However, the UKS also
has some synchronous methods which can't be forwarded to an async mojo
method. To solve this issue, the methods in the mojo service will also
return additional information that can be cached in the network process
and used to implement the synchronous part of the UKS interface.

So for example when creating or generating keys the mojo proxy won't
simply return the new key, but also metadata about what algorithm it is
using.

This CL finishes work started in CL 7079143, by now implementing the
Sign method.

Bug: 443932320, 452300505, 455518368
Change-Id: Ic9b87e3e0ab8a5635fdf64c303bca99276dadd91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7086090
Commit-Queue: Konstantin Andrikopoulos <andrikopoulos@google.com>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547932}