Disable sanitizer in gn instead of passing around an ignorelist
The array-bounds sanitizer should not be applied in
third_party/wuffs as it causes a performance degradation that shows
up in some benchmarks otherwise.

This was achieved by adding third_party/wuffs to a global ignorelist
that is passed around to clang everywhere, not only in wuffs where
needed.

The ignorelist breaks distributed builds using icecc, so the ignorelist
was disabled for icecc builds. So without the ignorelist icecc will work
again, but at the cost of the performance degradation coming back.

We can disable the sanitizer in wuffs by just removing the sanitizer
config in third_party/wuffs/BUILD.gn, which removes the need for a
global ignorelist.

This commit disables the sanitizer by removing the config in wuffs and
it removes the no longer needed ignorelist. We then avoid passing around
a global ignorelist everywhere and icecc will work again.

Bug: 457303640, 454399366
Change-Id: Ie2c42df2ee87c0a980bc8be8b85abc0b5cba6d00
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7172338
Reviewed-by: Gyuyoung Kim <gyuyoung@igalia.com>
Commit-Queue: Olle Liljenzin <ollel@opera.com>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547966}