[BNPL][Clank][Part2] Handle user clicking BNPL chip before AE completes
This CL handles the user flow for when the BNPL chip is clicked before
the transaction amount has been extracted.

When a user clicks the chip, Java calls the native side to show a
progress screen, as amount extraction is not yet complete. Once amount
extraction finishes and the amount is valid, the native side calls back
to Java with the BnplIssuerContext to display the BNPL issuer
selection screen.

Handling for invalid amounts will be addressed in a subsequent CL.

Video:
http://screencast/cast/NTc5MTg2NTE1OTgxMTA3Mnw4Mzk3N2ZkMC1hOQ

Design: go/autofill-clank-bnpl

Bug: b:438784412, 430575808
Change-Id: Ie8e520b0c77f85ca375b46ba400caa5207e7885d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7123178
Reviewed-by: Vinny Persky <vinnypersky@google.com>
Commit-Queue: Viplav Kadam <viplavkadam@google.com>
Reviewed-by: Timofey Chudakov <tchudakov@google.com>
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547972}