In wpt test, use performance.getEntriesByType for better precision.
We've not been able to figure out why on wpt.fyi, the
assert_equals for the font performance resource entry,
which uses performance.getEntriesByName
suggests that there are two entries, rather than one.
It's not reproducible on Chromium's try-bots, locally,
or even when creating a pending commit on github
for the web platform tests repo and using its try-bots.

So I'd like to make some reasonable change to this
file and see if it clears things up. :-)

Change-Id: I2e437b87c8a8cad8488828c8fe7bd3ca555b89d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7170721
Reviewed-by: Scott Haseley <shaseley@chromium.org>
Commit-Queue: Johannes Henkel <johannes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1547982}