viz: Fix visual artifacts due to resizing root render pass with DComp
crrev.com/1426453006 introduced an optimization that causes visual
artifacts with DComp.

During resizes, the browser compositor will often get at least one frame
behind the resizes that come from Windows. The existing optimization
resizes those frames to match the viewport size so that a swap is
possible.

When that resize makes the frame larger, it adds a gutter because the
frame's root render pass doesn't draw in the area that gets added.

With DComp, that gutter shows stale pixels. Whatever was previously
drawn to these areas of the output surface will become visible.

With viewporter, it is cheap to resize the viewport, so this patch
changes the logic to resize the viewport instead of the root render
pass. That eliminates the gutter and the stale pixels. It also makes
rendering faster because fewer pixels need to be updated.

The `resize_based_on_root_surface` capability causes the viewport size
to be set to the root render pass output rect in
`RootCompositorFrameSinkImpl::SubmitCompositorFrame`.

We put the logic behind a feature flag kill switch in case it causes
unexpected problems.

Interestingly, the original optimization is disabled on macOS because of
resize jank issues. See crrev.com/c/1069037.

Note: This does not fix ALL visual artifacts on DComp. For that, see
crrev.com/c/7115438.

---

Note: I cannot reproduce the issue with delegated compositing.

Delegated compositing is currently enabled on developer builds via a
field trial. To reproduce the issue and verify this fix, pass
`--disable-features=DelegatedCompositing` to your developer build or use
a branded Google Chrome build.

---

My work on this patch was funded by Plasticity.

---

Bug: 457463689
Change-Id: I7e00eadb2a1c3b44d64939d1a870d89befee158b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7156576
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Reviewed-by: Michael Tang <tangm@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1547987}