[BNPL][Clank][Part3] Handle user clicking BNPL chip before AE completes
This CL handles the user flow for when the BNPL chip is clicked before
the transaction amount has been extracted.

When a user clicks the chip, Java calls the native side to display a
progress screen, as amount extraction is not yet complete. Once
amount extraction finishes and the amount is not supported by any
issuer, the native side calls back to Java with the BnplIssuerContext
to display the BNPL issuer selection screen with all issuers grayed
out. If the amount is not present due to a timeout, an error screen is
shown.

Handling for graying out the BNPL chip on the home screen will
be addressed in a subsequent CL.

Video:
Amount Extraction returns with valid/invalid amount:
http://screencast/cast/NTg2MDUyMzE2NDgyNzY0OHw2MTNlZDVhZC1hOA
Amount Extraction returns with timeout:
http://screencast/cast/NjU3OTE2NzY5NTk5NDg4MHwyNGM2ZTFhNC1mMg

Design: go/autofill-clank-bnpl

Bug: b:438784412, 430575808
Change-Id: I07f4797ee120bcce302daba5f1ea85d3f67f0e5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7154302
Reviewed-by: Vinny Persky <vinnypersky@google.com>
Reviewed-by: Timofey Chudakov <tchudakov@google.com>
Commit-Queue: Viplav Kadam <viplavkadam@google.com>
Cr-Commit-Position: refs/heads/main@{#1548025}