[ios] Do a plain base copy of the registrar before ios/web migration
Do the minimal amount of changes to branch out the registrar from
autofill/ to ios/web

We did the following:
* Rename files (child => remote)
* Remove header includes that violated DEPS. Those will be added
  back in the follow up CL that does the real code changes for
  the migration.
* Put copyrights year to 2025 for the copies.
* Change the header precompile guards

Code changes for the migration are don in
https://chromium-review.googlesource.com/c/chromium/src/+/7110219
(follow-up CL).

The idea is to have a base to diff from so not all the migrated
code is seen as new code when reviewing the migration.

Gerrit can't track file branches.

Skipping presubmits to not have to make a build file for
the headers.

BYPASS_RECITATION_REASON=Plain copy of files in chromium, no AI used
NOPRESUBMIT=true

Change-Id: I1bde07b44f6afdc7cfec8056f62faaebb481120b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7168447
Commit-Queue: Vincent Boisselle <vincb@google.com>
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1548059}