net/disk_cache/sql: Avoid unnecessary SQL query on truncation to zero
When WriteEntry() is called with truncate=true and an end position of 0,
it results in a call to TrimOverlappingBlobs. This function would then
execute a SQL query (kTrimOverlappingBlobs_SelectOverlapping) to find
and trim blobs that overlap with the write range.

This SELECT query is unnecessary in this specific scenario, as no blobs
can exist before offset 0. This was discovered while analyzing traces.

This change adds a check at the beginning of TrimOverlappingBlobs to
return early if the entry is being truncated to zero length, thus
avoiding the unnecessary database call.

Fixed: 462259321
Change-Id: Ieed611eae4aeb9d36a404ce0c10b7681eff61ac2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7172086
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1548077}