Reland: Animate compact mode expand/contract
This is accomplished with the transition framework. LBC is notified
right before layout changes corresponding to compact mode state changes
are made, allowing it to capture values at the right time. The location
bar itself and the "+" button are animated.

In order to animate contraction correctly, the Toolbar needs to remain
larger until the animation ends to avoid clipping the LocationBar.
LocationBarEmbedder allows for the toolbar's height to temporarily fixed
without exposing the view directly.

In order to provide consistent notifications re: compact mode state, we
need to create a single source of truth for compactness, extracting the
view binder's control of this property out to the mediator.

Reland changes: Early return if the add button is not present, don't
observe if flag is disabled.

Original CL: https://chromium-review.googlesource.com/c/chromium/src/+/7171390

Bug: TBD
Change-Id: Ic892cf746783cc19f36f2e1d35fec92c6239fe6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7180290
Reviewed-by: Sky Malice <skym@chromium.org>
Reviewed-by: Tomasz Wiszkowski <ender@google.com>
Commit-Queue: Patrick Noland <pnoland@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1548101}