Change file handle id from size_t to uint64_t in FontDataService
This is equivalent, but better reflects the type in the mojom file.

Also considered was base::IdType, but the lack of a well-fitting type
for the IdType's template argument, and the limited usage of the ID
(which is only created to pass over mojom) made this option less
attractive.

Bug: 461472567
Change-Id: I277245a9185980f051f614cf9b4dd9c101e96838
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7163416
Commit-Queue: Anthony Vallée-Dubois <anthonyvd@chromium.org>
Reviewed-by: Joe Mason <joenotcharles@google.com>
Cr-Commit-Position: refs/heads/main@{#1548156}