[Connection-Allowlist] Adding network restrictions id to URLLoaderFactory
This CL is the first one in the series that moves the functionality
of Connection Allowlist prototype to check the enforcements in the content/browser and network service layers. The idea is to reuse the
infrastructure for network revocation that is already implemented in
these layers for fenced frames disableUntrustedNetwork() API (see
NetworkContext:: IsNetworkForNonceAndUrlAllowed()).
For reference, connection allowlists's prototype started in the chain
of CLs beginning at https://chromium-review.googlesource.com/c/chromium/src/+/6985001.

This one does the following:
1. Creates an identifier in a given frame's URLLoaderFactory so any
requests that go through this factory can be checked to apply network
restrictions, if applicable. The check itself will be in the follow up.
2. It reuses the value of DocumentAssociatedData::token() for this id
since it is already uniquely created for each document.

Tests: This CL only assigns the id in the URLLoaderFactory, which will
be a path exercised by many existing tests. The next CL will actually
test the network restrictions being checked.

TODOs: This Cl only assigns the network_restriction_id for a frame's
subresource URLLoaderFactory. The other cases will be done in follow
ups. Those cases will also use the DocumentAssociatedData::token() of
the associated frame, wherever possible.

Bug: 447954811

Change-Id: I483309f6632150dbc2ad73cfe2ac7dcb4d147dfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7082870
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Shivani Sharma <shivanisha@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1548189}