[Connection-Allowlist] Enforcing fetch restrictions in the network service
Explainer: https://github.com/WICG/connection-allowlists

This CL is the second one in the series that implements the
functionality of Connection Allowlist prototype in the content/browser
and network service layers. This one does the following:
1. Introduces a network restriction id in the NavigationRequest and also
updates it in the DocumentAssociatedData, so it is cleared from the
network service when the document goes away.
The id gets propagated to the network service map that keeps the
nonces that have network restrictions applied. It is only passed if the
connection allowlist header has a non-empty enforced list. It is removed
in DocumentAssociatedData's destructor or if the navigation fails.
2. Updates `network_revocation_nonces_` in both StoragePartition and
NetworkContext to include the connection allowlists.
3. The network restriction and connection allowlists are checked in
every outgoing request from the subresource URLLoaderFactory.
TODO for follow up CLs: Currently the connection allowlists check is
based on origin matching but will be changed to URLPattern in follow up
CLs.
4. Finally, this CL removes the enforcement from blink.

Tests:
Added browser test RenderFrameHostImplConnectionAllowlistBrowserTest*
Existing WPT :
virtual/connection-allowlist/external/wpt/connection-allowlist/tentative/

TODOs:
This CL only enforces the network restrictions on subresource requests.
Other network end points will be handled in follow ups.

Bug: 447954811

Change-Id: I605323e054a70bf1e77f3e42384cb4f19b3feb3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7085134
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Shivani Sharma <shivanisha@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1548192}