[contextual_tasks] Added attached tabs menu dropdown
This CL adds the UI for attached tabs. A placeholder icon was
added which when clicked will reveal the list of tabs attached to
the task. The tab item when clicked from the dropdown will result
in a tab switch. If the tab isn't a open tab, the browser will
create a new tab with the given URL.

Bug: 460614856
Change-Id: Id6e17e8b265018c3dbfaf5f376621beeb82b0350
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7154021
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Shakti Sahu <shaktisahu@chromium.org>
Reviewed-by: Mickey Burks <mickeyburks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1548207}